Extract abstract class for lesson progress
